refactor: 랜덤정수 적용 테스트를 위해, 정수 생성기를 인터페이스화하고, 테스트용 고정 자연수 생성기 추가
