+++ src/main/java/racingcar/domain/CarRacing.java
@@ -4,11 +4,12 @@ package racingcar.domain;
 import racingcar.domain.output.RacingPrinter;
 
 public class CarRacing {
-    private final ApplicatorRandomNumberToCar applicatorRandomNumberToCar;
+    private final ApplicatorNumbersToCar applicatorRandomNumberToCar;
     private final RacingPrinter racingPrinter;
 
     public CarRacing(CarRepository carRepository) {
-        this.applicatorRandomNumberToCar = new ApplicatorRandomNumberToCar(carRepository);
+        this.applicatorRandomNumberToCar
+            = new ApplicatorNumbersToCar(carRepository, new RandomNumberGenerator());
         this.racingPrinter = new RacingPrinter(carRepository);
     }
