+++ src/main/java/racingcar/domain/NumberGenerator.java
@@ -1,0 +1,5 @@
+package racingcar.domain;
+
+public interface NumberGenerator {
+    int generate();
+}
