+++ src/test/java/racingcar/domain/FixedNumberGenerator.java
@@ -1,0 +1,10 @@
+package racingcar.domain;
+
+
+public class FixedNumberGenerator implements NumberGenerator {
+    private final int FIXED_NUMBER = 5;
+
+    public int generate() {
+        return FIXED_NUMBER;
+    }
+}
